requirements.txt
torch
numpy
PyYAML
setuptools
torchinfo
usd-core
